Revert checkout function
