+++ payments-ui-core/api/payments-ui-core.api
@@ -201,9 +201,6 @@ public final class com/stripe/android/ui/core/elements/SimpleTextSpec$Companion 
 public final class com/stripe/android/ui/core/elements/StaticTextElementUIKt {
 }
 
-public final class com/stripe/android/ui/core/elements/menu/CheckboxKt {
-}
-
 public final class com/stripe/android/ui/core/forms/AffirmSpecKt {
 	public static final fun getAffirmForm ()Lcom/stripe/android/ui/core/elements/LayoutSpec;
 	public static final fun getAffirmParamKey ()Ljava/util/Map;
